requirements.txt
python-ldap==3.4.3
pyyaml==6.0
jinja2==3.1.2
requests==2.31.0
smtplib2==0.2.1
schedule==1.2.0
pymsteams==0.2.2
